+++ acceptance-tests/src/main/java/com/acme/Person.java
@@ -1,0 +1,7 @@
+package com.acme;
+
+import nativeimage.Reflection;
+
+@Reflection(constructors = true)
+public class Person {
+}
