+++ acceptance-tests/src/main/java/com/acme/subpackage/UpperCass.java
@@ -1,0 +1,6 @@
+package com.acme.subpackage;
+
+public class UpperCass {
+	public static class InnerClass {
+	}
+}
