+++ acceptance-tests/src/test/resources/reflection-config-generation-test/001.json
@@ -9,6 +9,21 @@
     "allDeclaredFields" : true
   },
   {
+    "name" : "com.acme.Person",
+    "allDeclaredConstructors" : false,
+    "allPublicConstructors" : false,
+    "allDeclaredMethods" : false,
+    "allPublicMethods" : false,
+    "allPublicFields" : false,
+    "allDeclaredFields" : false,
+    "methods" : [
+      {
+        "name" : "<init>",
+        "parameterTypes" : [ ]
+      }
+    ]
+  },
+  {
     "name" : "com.acme.Fruit",
     "allDeclaredConstructors" : true,
     "allPublicConstructors" : true,
@@ -28,6 +43,15 @@
   },
   {
     "name" : "com.acme.subpackage.Color",
+    "allDeclaredConstructors" : false,
+    "allPublicConstructors" : false,
+    "allDeclaredMethods" : false,
+    "allPublicMethods" : false,
+    "allPublicFields" : false,
+    "allDeclaredFields" : true
+  },
+  {
+    "name" : "com.acme.subpackage.UpperCass",
     "allDeclaredConstructors" : false,
     "allPublicConstructors" : false,
     "allDeclaredMethods" : false,
