+++ src/main/java/nativeimage/core/ReflectionConfigAppender.java
@@ -1,5 +1,7 @@
 package nativeimage.core;
 
+import nativeimage.core.domain.ReflectionConfig;
+
 public interface ReflectionConfigAppender {
 	void append(ReflectionConfig reflectionConfig);
 }
